requirements.txt
PyYAML
requests
pdf2docx
argparse